backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]
SQLAlchemy==2.0.25
psycopg2-binary
alembic
python-dotenv
pydantic>=2.5.0
pydantic-settings
redis
passlib==1.7.4
bcrypt==4.1.2
python-jose[cryptography]
authlib
httpx
itsdangerous
